+++ Tests/Functional/Domain/Repository/Event/EventRepositoryTest.php
@@ -84,7 +84,7 @@ final class EventRepositoryTest extends FunctionalTestCase
      */
     public function mapsSingleEventWithAllFields(): void
     {
-        $this->importCSVDataSet(__DIR__ . '/Fixtures/SingleEventWithAllFields.csv');
+        $this->importCSVDataSet(__DIR__ . '/Fixtures/propertyMapping/SingleEventWithAllFields.csv');
 
         $result = $this->subject->findByUid(1);
 
@@ -117,6 +117,7 @@ final class EventRepositoryTest extends FunctionalTestCase
         self::assertSame('https://webinar.example.com/', $result->getWebinarUrl());
         self::assertSame('Some more text for the attendees', $result->getAdditionalEmailText());
         self::assertSame('conference room 2', $result->getRoom());
+        self::assertSame('jousting/1', $result->getSlug());
     }
 
     /**
@@ -312,7 +313,7 @@ final class EventRepositoryTest extends FunctionalTestCase
      */
     public function mapsEmptyEventTypeAssociationForSingleEventAsNull(): void
     {
-        $this->importCSVDataSet(__DIR__ . '/Fixtures/SingleEventWithAllFields.csv');
+        $this->importCSVDataSet(__DIR__ . '/Fixtures/propertyMapping/SingleEventWithAllFields.csv');
 
         $result = $this->subject->findByUid(1);
         self::assertInstanceOf(SingleEvent::class, $result);
@@ -778,7 +779,7 @@ final class EventRepositoryTest extends FunctionalTestCase
      */
     public function updateRegistrationCounterCacheForNoRegistrationsSetsCounterCacheAtZero(): void
     {
-        $this->importCSVDataSet(__DIR__ . '/Fixtures/SingleEventWithAllFields.csv');
+        $this->importCSVDataSet(__DIR__ . '/Fixtures/propertyMapping/SingleEventWithAllFields.csv');
         $event = $this->subject->findByUid(1);
 
         $this->subject->updateRegistrationCounterCache($event);
@@ -1553,7 +1554,7 @@ final class EventRepositoryTest extends FunctionalTestCase
      */
     public function enrichWithRawDataAddsRawDataToEvents(): void
     {
-        $this->importCSVDataSet(__DIR__ . '/Fixtures/SingleEventWithAllFields.csv');
+        $this->importCSVDataSet(__DIR__ . '/Fixtures/propertyMapping/SingleEventWithAllFields.csv');
         $event = $this->subject->findByUid(1);
         self::assertInstanceOf(SingleEvent::class, $event);
         $events = [$event];