+++ Tests/Unit/Domain/Model/Event/EventTopicTest.php
@@ -589,4 +589,23 @@ final class EventTopicTest extends UnitTestCase
     {
         self::assertFalse($this->subject->isRegistrationPossibleByDate());
     }
+
+    /**
+     * @test
+     */
+    public function getSlugInitiallyReturnsEmptyString(): void
+    {
+        self::assertSame('', $this->subject->getSlug());
+    }
+
+    /**
+     * @test
+     */
+    public function setSlugSetsSlug(): void
+    {
+        $value = 'best-thing-ever/3';
+        $this->subject->setSlug($value);
+
+        self::assertSame($value, $this->subject->getSlug());
+    }
 }